chore: update project structure by removing favicon and adjusting build scripts

+++ packages/electron/index.js
@@ -1,43 +1,72 @@
 const { app, BrowserWindow, Menu } = require("electron");
-const path = require("path");
 const { spawn } = require("child_process");
+const path = require("path");
 
 let window;
 let serverProcess;
 
-app.whenReady().then(() => {
+const gotTheLock = app.requestSingleInstanceLock();
 
-  serverProcess = spawn("node", [path.join(__dirname, "../server/index.js")], {
-    stdio: "ignore",
-    detached: true
-  });
-  serverProcess.unref();
+if (!gotTheLock) {
+  app.quit();
+} 
+else {
+  app.whenReady().then(() => {
+    Menu.setApplicationMenu(null);
 
-  Menu.setApplicationMenu(null);
-
-  window = new BrowserWindow({
-    width: 1280,
-    height: 720,
-    minWidth: 1280,
-    minHeight: 720,
-    title: "Siete Colores",
-    icon: "/assets/favicon.ico",
-    webPreferences: {
-      nodeIntegration: false,
-      contextIsolation: true
+    if (app.isPackaged) {
+      serverProcess = spawn(
+        process.execPath, 
+        [path.join(__dirname, "server", "app.js")], 
+        { stdio: "ignore", detached: true }
+      );
+    } 
+    else {
+      serverProcess = spawn("node", [path.join(__dirname, "../server/src/app.ts")], {
+        stdio: "ignore",
+        detached: true
+      });
     }
+
+    serverProcess.unref();
+
+    window = new BrowserWindow({
+      width: 1280,
+      height: 720,
+      minWidth: 1280,
+      minHeight: 720,
+      title: "Siete Colores",
+      icon: path.join(__dirname, "assets", "icon.svg"),
+      webPreferences: {
+        nodeIntegration: false,
+        contextIsolation: true
+      }
+    });
+
+    if (app.isPackaged) {
+      const indexFilePath = path.join(__dirname, "client", "index.html");
+      window.loadURL(`file://${indexFilePath}`);
+    } 
+    else {
+      setTimeout(() => window.loadURL("http://localhost:5173"), 5000);
+    }
+
+    app.on('second-instance', () => {
+      if (window) {
+        if (window.isMinimized()) window.restore();
+        window.focus();
+      }
+    });
+
+    window.on("closed", () => {
+      window = null;
+      if (serverProcess) serverProcess.kill();
+    });
+
   });
 
-  setTimeout(() => window.loadURL("http://localhost:5173"), 5000);
-
-  window.on("closed", () => {
-    window = null;
+  app.on("window-all-closed", () => {
     if (serverProcess) serverProcess.kill();
+    if (process.platform !== "darwin") app.quit();
   });
-
-});
-
-app.on("window-all-closed", () => {
-  if (serverProcess) serverProcess.kill();
-  if (process.platform !== "darwin") app.quit();
-});
+}
